+++ app/react-native.config.js
@@ -4,4 +4,14 @@ module.exports = {
         android:{}
     },
     assets:['./src/Assets/Fonts/'],
+    dependencies: {
+      'react-native-video': {
+        platforms: {
+          android: {
+            sourceDir: '../node_modules/react-native-video/android-exoplayer',
+          },
+        },
+      },
+    },
 }
+