+++ app/android/CHANGELOG.md
@@ -7,6 +7,27 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [1.0.6 (version code 7)] - 2022-08-23Z
+
+### Changed
+
+- Vimeo MP4 video file is now part of the project cache 
+
+### Added
+
+- Video component to display the Vimeo video file in the app
+
+## [1.0.5 (version code 6)] - 2022-08-22Z
+
+### Changed
+
+- API results for events now give the event duration as a number of minutes
+
+### Added
+
+- Upcoming events quick search by date (today, this week, this month)
+- Upcoming events search using calendar picker now fully working
+
 ## [1.0.3 (version code 5)] - 2022-08-12Z
 
 ### Changed